+++ saveFile.txt
@@ -1,0 +1,5 @@
+T 1 read book
+D 1 eat dinner /by 7pm tonight
+E 1 meet RF /from 630pm /to 7pm
+E 1 go sleep /from 121 /to 213
+T 0 code
