+++ src/main/java/Event.java
@@ -11,4 +11,8 @@ public class Event extends Task {
     public String toString() {
         return String.format("[E][%s] %s (from: %s to: %s)", getStatusIcon(), this.description, this.start, this.end);
     }
+    @Override
+    public String getSaveLine() {
+        return "E " + (this.isDone ? "1 " : "0 ") + this.description + " /from " + this.start + " /to " + this.end + "\n";
+    }
 }
